+++ CHANGELOG.md
@@ -3,7 +3,14 @@
 ### Added
 - Sonic Screwdrivers now power gravity and oxygen generators.
 - 10th Doctor's Sonic Screwdriver.
+- Steel. Including raw, ingot and blocks.
 
-#### Fixed
+### Fixed
 - Fixed TARDIS gravity generator not working.
 - Added a rendering fix for compatibility with the skin shuffle mod.
+
+### Planned
+- Cybermen.
+- Cyberships.
+- Cybermats.
+- Cyberconversion.